Move credential_loop out of Base
